feat(coupons): Revamped the structure to be modular
