+++ internal/service/subscription.go
@@ -43,8 +43,6 @@ type SubscriptionService interface {
 
 	// Coupon-related methods
 	ApplyCouponsToSubscription(ctx context.Context, subscriptionID string, couponIDs []string) error
-	GetSubscriptionCouponAssociations(ctx context.Context, subscriptionID string) ([]*dto.CouponAssociationResponse, error)
-	RemoveCouponFromSubscription(ctx context.Context, subscriptionID string, couponID string) error
 }
 
 type subscriptionService struct {
@@ -452,8 +450,8 @@ func (s *subscriptionService) GetSubscription(ctx context.Context, id string) (*
 	}
 
 	// expand coupon associations
-	couponService := NewCouponService(s.ServiceParams)
-	couponAssociations, err := couponService.GetCouponAssociationsBySubscription(ctx, id)
+	couponAssociationService := NewCouponAssociationService(s.ServiceParams)
+	couponAssociations, err := couponAssociationService.GetCouponAssociationsBySubscription(ctx, id)
 	if err != nil {
 		s.Logger.Errorw("failed to get coupon associations for subscription",
 			"subscription_id", id,
@@ -2536,10 +2534,10 @@ func (s *subscriptionService) ApplyCouponsToSubscription(ctx context.Context, su
 		"coupon_count", len(couponIDs))
 
 	// Create coupon service instance
-	couponService := NewCouponService(s.ServiceParams)
+	couponAssociationService := NewCouponAssociationService(s.ServiceParams)
 
 	// Apply coupons to subscription
-	_, err := couponService.ApplyCouponToSubscription(ctx, couponIDs, subscriptionID)
+	err := couponAssociationService.ApplyCouponToSubscription(ctx, couponIDs, subscriptionID)
 	if err != nil {
 		return ierr.WithError(err).
 			WithHint("Failed to apply coupons to subscription").
@@ -2555,56 +2553,4 @@ func (s *subscriptionService) ApplyCouponsToSubscription(ctx context.Context, su
 		"coupon_count", len(couponIDs))
 
 	return nil
-}
-
-// GetSubscriptionCouponAssociations retrieves all coupon associations for a subscription
-func (s *subscriptionService) GetSubscriptionCouponAssociations(ctx context.Context, subscriptionID string) ([]*dto.CouponAssociationResponse, error) {
-	couponService := NewCouponService(s.ServiceParams)
-	return couponService.GetCouponAssociationsBySubscription(ctx, subscriptionID)
-}
-
-// RemoveCouponFromSubscription removes a coupon association from a subscription
-func (s *subscriptionService) RemoveCouponFromSubscription(ctx context.Context, subscriptionID string, couponID string) error {
-	couponService := NewCouponService(s.ServiceParams)
-
-	// Get coupon associations for this subscription
-	associations, err := couponService.GetCouponAssociationsBySubscription(ctx, subscriptionID)
-	if err != nil {
-		return ierr.WithError(err).
-			WithHint("Failed to get coupon associations").
-			Mark(ierr.ErrInternal)
-	}
-
-	// Find the association with the specified coupon ID
-	for _, association := range associations {
-		if association.CouponID == couponID {
-			// Delete the association
-			err = couponService.DeleteCouponAssociation(ctx, association.ID)
-			if err != nil {
-				return ierr.WithError(err).
-					WithHint("Failed to remove coupon from subscription").
-					WithReportableDetails(map[string]interface{}{
-						"subscription_id": subscriptionID,
-						"coupon_id":       couponID,
-						"association_id":  association.ID,
-					}).
-					Mark(ierr.ErrInternal)
-			}
-
-			s.Logger.Infow("successfully removed coupon from subscription",
-				"subscription_id", subscriptionID,
-				"coupon_id", couponID,
-				"association_id", association.ID)
-
-			return nil
-		}
-	}
-
-	return ierr.NewError("coupon not found in subscription").
-		WithHint("The specified coupon is not associated with this subscription").
-		WithReportableDetails(map[string]interface{}{
-			"subscription_id": subscriptionID,
-			"coupon_id":       couponID,
-		}).
-		Mark(ierr.ErrNotFound)
 }
